requirements.txt
streamlit
psycopg2-binary
supabase
python-dotenv
pandas